+++ Fossil/src/Fossil.h
@@ -7,8 +7,10 @@
 #include <fstream>
 #include <vector>
 
+#include "BuildDll.h"
 #include "Settings.h"
 #include "Log.h"
-#include "Game.h"
-#include "BuildDll.h"
 #include "EntryPoint.h"
+#include "Game.h"
+#include "GameState.h"
+#include "GameStateStack.h"