+++ Fossil/src/GameState.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "BuildDll.h"
+
+class FSL_API GameState
+{
+public:
+	virtual ~GameState();
+
+	virtual void InitializeState() = 0;
+
+	virtual void PauseState() = 0;
+	virtual void ResumeState() = 0;
+
+	virtual void UpdateState() = 0;
+	virtual void DisplayState() = 0;
+
+protected:
+	GameState();
+};